fetch/metadata/generated/window-location-sec-fetch-mode.https.sub.html

<!DOCTYPE html>
<!-- This test was procedurally generated. Please do not modify it directly. -->
<html lang="en">
  <meta charset="utf-8">
  <title>Fetch Metadata - Sec-Fetch-Mode</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <body>
  <script>
  'use strict';

  function induceRequest(url, navigate, userActivated) {
    const win = window.open();

    return new Promise((resolve) => {
        addEventListener('message', function(event) {
          if (event.source === win) {
            resolve();
          }
        });

        if (userActivated) {
          test_driver.bless('enable user activation', () => {
            navigate(win, url);
          });
        } else {
          navigate(win, url);
        }
      })
      .then(() => win.close());
  }

  const responseParams = {
    mime: 'text/html',
    body: `<script>opener.postMessage('done', '*')</${''}script>`
  };

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, [], responseParams);

    const navigate = (win, path) => {
      win.location = path;
    };
    return induceRequest(url, navigate, false)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'navigate');
        });
  }, 'location');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, [], responseParams);

    const navigate = (win, path) => {
      win.location.href = path;
    };
    return induceRequest(url, navigate, false)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'navigate');
        });
  }, 'location.href');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, [], responseParams);

    const navigate = (win, path) => {
      win.location.assign(path);
    };
    return induceRequest(url, navigate, false)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'navigate');
        });
  }, 'location.assign');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, [], responseParams);

    const navigate = (win, path) => {
      win.location.replace(path);
    };
    return induceRequest(url, navigate, false)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'navigate');
        });
  }, 'location.replace');
  </script>
  </body>
</html>
